documents-expolorer components breakdown
